FUSIO-450: Fix issues
* FUSIO-450 - fix not closed invite
* FUSIO-451 - fix knocks

+++ tests/sanity/tests/global.setup.ts
@@ -16,5 +16,7 @@
 import { wipeStats } from './stats'
 
 export default async function globalSetup (): Promise<void> {
-  await wipeStats()
+  if (process.env.STATS === '1' || process.env.STATS === 'true') {
+    await wipeStats()
+  }
 }
